Add news archives & singles pages and misc changes
Set up the backend like the WMS project with a FUST class, custom
post types, default single and archive templates, and added a concrete
implementation of the news post type. Also added a simple 404 page,
an alternative style for the home page about section and a fix for
the misalignment of the design splashes on the home page.