Implement guest membership
Added the ability for users to apply as guest member. This will send an
email to the association of which the user claims to be a part of, such
that the assocation can confirm this to FUST. Upon confirmation, admins
can create a new guest member account for the applicant on the revamped
FUST admin page.

Also seperated the welcome emails for guest members and normal members,
and added a `guest-member` role to automatically assign to guest member
accounts.